Conflicts:
      doc/admin-guide/files/records.yaml.en.rst

+++ src/iocore/cache/AggregateWriteBuffer.cc
@@ -65,6 +65,6 @@ AggregateWriteBuffer::flush(int fd, off_t write_pos) const
 void
 AggregateWriteBuffer::copy_from(char *dest, int offset, size_t nbytes) const
 {
-  ink_assert((offset + nbytes) <= (unsigned)this->_buffer_pos);
+  ink_assert((offset + nbytes) <= static_cast<size_t>(this->_buffer_pos));
   memcpy(dest, this->_buffer + offset, nbytes);
 }
